requirements.txt
httpx
pandas
beautifulsoup4
python-dotenv
fake-useragent
